+++ spec/app/models/person.rb
@@ -23,6 +23,7 @@ class Person
   field :security_code
   field :reading, :type => Object
   field :bson_id, :type => BSON::ObjectId
+  field :pattern, :type => Regexp
 
   index :age
   index :addresses